Updated firewall rules
Added more secure firewall rules, instead of using iptables we're going to use OpenWRT's built in firewall, as its shown to be more reliable. ** Please note ** Unfortunately we can't block all 10.0.0.* addresses due to them being used on the mesh for WLAN addresses. If a client is using a 10.0.0.* address we need to add a custom rule for them (Unless they're running a custom network setup this shouldn’t be an issue) - Comcast Customers are covered by these rules as well (We block access to the 10.1.10.0/24 range that Comcast provides its customers)